Changed the parameter configuration collection into a name value collection
