--- src/DotNetToolkit.Repository/Internal/ConfigFile/CachingProviderElement.cs
@@ -1,12 +1,0 @@
-﻿#if !NETSTANDARD1_3
-
-namespace DotNetToolkit.Repository.Internal.ConfigFile
-{
-    using Configuration.Caching;
-
-    internal class CachingProviderElement : TypedConfigurationElementBase<ICacheProvider>
-    {
-    }
-}
-
-#endif
