--- src/DotNetToolkit.Repository/Internal/ConfigFile/LoggingProviderElement.cs
@@ -1,12 +1,0 @@
-﻿#if !NETSTANDARD1_3
-
-namespace DotNetToolkit.Repository.Internal.ConfigFile
-{
-    using Configuration.Logging;
-
-    internal class LoggingProviderElement : TypedConfigurationElementBase<ILoggerProvider>
-    {
-    }
-}
-
-#endif
