--- src/DotNetToolkit.Repository/Internal/ConfigFile/MappingProviderElement.cs
@@ -1,12 +1,0 @@
-﻿#if !NETSTANDARD1_3
-
-namespace DotNetToolkit.Repository.Internal.ConfigFile
-{
-    using Configuration.Mapper;
-
-    internal class MappingProviderElement : TypedConfigurationElementBase<IMapperProvider>
-    {
-    }
-}
-
-#endif
