--- src/DotNetToolkit.Repository/Internal/ConfigFile/RepositoryContextFactoryElement.cs
@@ -1,12 +1,0 @@
-﻿#if !NETSTANDARD1_3
-
-namespace DotNetToolkit.Repository.Internal.ConfigFile
-{
-    using Factories;
-
-    internal class RepositoryContextFactoryElement : TypedConfigurationElementBase<IRepositoryContextFactory>
-    {
-    }
-}
-
-#endif
