--- src/DotNetToolkit.Repository/Internal/ConfigFile/RepositoryInterceptorElement.cs
@@ -1,12 +1,0 @@
-﻿#if !NETSTANDARD1_3
-
-namespace DotNetToolkit.Repository.Internal.ConfigFile
-{
-    using Configuration.Interceptors;
-
-    internal class RepositoryInterceptorElement : TypedConfigurationElementBase<IRepositoryInterceptor>
-    {
-    }
-}
-
-#endif
