+++ test/DotNetToolkit.Repository.Test/App.config
@@ -6,27 +6,27 @@
   </configSections>
   <repository>
     <defaultContextFactory type="DotNetToolkit.Repository.InMemory.Internal.InMemoryRepositoryContextFactory, DotNetToolkit.Repository.InMemory">
-      <keyValues>
-        <add key="ignoreTransactionWarning" value="True" />
-        <add key="databaseName" value="__InMemoryDatabaseName__" />
-      </keyValues>
+      <parameters>
+        <parameter name="ignoreTransactionWarning" value="True" />
+        <parameter name="databaseName" value="__InMemoryDatabaseName__" />
+      </parameters>
     </defaultContextFactory>
     <loggingProvider type="DotNetToolkit.Repository.Configuration.Logging.ConsoleLoggerProvider, DotNetToolkit.Repository">
       <parameters>
-        <parameter value="Debug" type="DotNetToolkit.Repository.Configuration.Logging.LogLevel, DotNetToolkit.Repository" />
+        <parameter name="minLogLevel" value="Debug" />
       </parameters>
     </loggingProvider>
     <cachingProvider type="DotNetToolkit.Repository.Extensions.Microsoft.Caching.Memory.InMemoryCacheProvider, DotNetToolkit.Repository.Extensions.Microsoft.Caching.Memory">
-      <keyValues>
-        <add key="expiry" value="00:00:30" />
-      </keyValues>
+      <parameters>
+        <parameter name="expiry" value="00:00:30" />
+      </parameters>
     </cachingProvider>
     <mappingProvider type="DotNetToolkit.Repository.Configuration.Mapper.MapperProvider, DotNetToolkit.Repository" />
     <interceptors>
       <interceptor type="DotNetToolkit.Repository.Test.Data.TestRepositoryInterceptor, DotNetToolkit.Repository.Test">
         <parameters>
-          <parameter value="random param" />
-          <parameter value="True" type="System.Boolean" />
+          <parameter name="p1" value="random param" />
+          <parameter name="p2" value="True" />
         </parameters>
       </interceptor>
     </interceptors>
