+++ test/DotNetToolkit.Repository.Test/appsettings.json
@@ -2,25 +2,22 @@
   "repository": {
     "defaultContextFactory": {
       "type": "DotNetToolkit.Repository.InMemory.Internal.InMemoryRepositoryContextFactory, DotNetToolkit.Repository.InMemory",
-      "keyValues": {
-        "ignoreTransactionWarning": "True",
+      "parameters": {
+        "ignoreTransactionWarning": true,
         "databaseName": "__InMemoryDatabaseName__"
       }
     },
     "loggingProvider": {
       "type": "DotNetToolkit.Repository.Configuration.Logging.ConsoleLoggerProvider, DotNetToolkit.Repository",
-      "parameters": [
-        {
-          "type": "DotNetToolkit.Repository.Configuration.Logging.LogLevel, DotNetToolkit.Repository",
-          "value": "Debug"
-        }
-      ]
+      "parameters": {
+        "minLogLevel": "Debug"
+      }
     },
     "cachingProvider": {
       "type": "DotNetToolkit.Repository.Extensions.Microsoft.Caching.Memory.InMemoryCacheProvider, DotNetToolkit.Repository.Extensions.Microsoft.Caching.Memory",
-      "keyValues": {
+      "parameters": {
         "expiry": "00:00:30"
-      } 
+      }
     },
     "mappingProvider": {
       "type": "DotNetToolkit.Repository.Configuration.Mapper.MapperProvider, DotNetToolkit.Repository"
@@ -28,16 +25,10 @@
     "interceptors": [
       {
         "type": "DotNetToolkit.Repository.Test.Data.TestRepositoryInterceptor, DotNetToolkit.Repository.Test",
-        "parameters": [
-          {
-            "type": "System.String",
-            "value": "random param"
-          },
-          {
-            "type": "System.Boolean",
-            "value": "True"
-          }
-        ]
+        "parameters": {
+          "p1": "random param",
+          "p2": true
+        }
       }
     ]
   }
